python/model_zoom/requirements.txt
minio~=7.1.13
requests~=2.28.1
opencv-python~=4.8.0.74
numpy~=1.23.4
onnxruntime~=1.14.0
Pillow~=9.5.0
setuptools~=59.8.0
gevent~=21.1.2
Flask~=2.2.2
pypinyin~=0.44.0
jieba~=0.42.1
jieba3k~=0.35.1
torch~=1.13.0
transformers~=4.25.1
synonyms~=3.18.0
snownlp~=0.12.3
joblib~=1.2.0
pandas~=1.3.5
textrank4zh~=0.3
nltk~=3.5
transliterate~=1.10.2
scipy~=1.9.3
xpinyin~=0.7.6
librosa~=0.9.2
inflect~=5.3.0
Unidecode~=1.3.2
matplotlib~=3.2.2
dlib~=19.24.0
loguru~=0.5.3
tqdm~=4.64.1
pyclipper~=1.2.1
Shapely~=1.7.1
clip~=1.0
scikit-image~=0.18.1
vcam~=1.0